Fill organization and starter code for the frontend
Created Pages folder for the website pages and moved the Entrepreneur file into it. Also create Authentication (Sign-in and Sign-up) file and Profile file. Also create Components folder which holds components and logic for navigating around the application, like menus, links, and route setups
