+++ package.json
@@ -11,6 +11,7 @@
     "detect-port-alt": "^1.1.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.0",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
