+++ src/Components/Navigation.css
@@ -1,0 +1,28 @@
+nav {
+    background-color: #333; /* Dark background color */
+    padding: 10px;
+  }
+  
+  ul {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+  }
+  
+  li {
+    display: inline;
+    margin-right: 10px;
+  }
+  
+  /* Link styles */
+  a {
+    text-decoration: none;
+    color: #fff; /* White color */
+    padding: 5px 10px;
+    border-radius: 3px;
+  }
+  
+  a:hover {
+    background-color: #555; /* Darker background color on hover */
+  }
+
