+++ src/Pages/Profile/Profile.js
@@ -1,0 +1,56 @@
+// This is the starter code for the profile page
+
+import React from 'react';
+import './Profile.css'; // Import the CSS file for styling
+
+const Profile = () => {
+    return (
+        <div className="profile-container">
+            <div className="summary-dashboard">
+                <h2>Summary Dashboard</h2>
+                {/* Overview Metrics */}
+                <div className="overview-metrics">
+                    {/* Metrics components */}
+                </div>
+                {/* Recent Activity */}
+                <div className="recent-activity">
+                    {/* Recent activity components */}
+                </div>
+            </div>
+            <div className="tabs-sections">
+                <h2>Tabs or Sections</h2>
+                {/* Tabs or sections components */}
+            </div>
+            <div className="interaction-history">
+                <h2>Interaction History</h2>
+                {/* Interaction history components */}
+            </div>
+            <div className="notes-comments">
+                <h2>Notes and Comments Section</h2>
+                {/* Notes and comments components */}
+            </div>
+            <div className="customer-segmentation">
+                <h2>Customer Segmentation</h2>
+                {/* Segmentation information components */}
+            </div>
+            <div className="account-information">
+                <h2>Account Information</h2>
+                {/* Account information components */}
+            </div>
+            <div className="action-buttons">
+                <h2>Action Buttons</h2>
+                {/* Action buttons components */}
+            </div>
+            <div className="search-filter-options">
+                <h2>Search and Filter Options</h2>
+                {/* Search and filter options components */}
+            </div>
+            <div className="security-privacy-controls">
+                <h2>Security and Privacy Controls</h2>
+                {/* Security and privacy controls components */}
+            </div>
+        </div>
+    );
+}
+
+export default Profile;
